Various updates
various changes to bring back the project to running again, as parcel seemed to have been a problem. Now switched to vite.
Changed the UpdateButton to a UpdateFormView. Now almost functional: had to change the backend to unblock the possibility of updating only 1 or two fields, yet it's still not working